+++ src/pyrefdev/config.py
@@ -115,6 +115,7 @@ _packages = [
     Package(pypi="networkx", index_url="https://networkx.org/documentation/stable/reference/index.html"),
     Package(pypi="ipython", index_url="https://ipython.readthedocs.io/en/stable/"),
     Package(pypi="httplib2", index_url="https://httplib2.readthedocs.io/en/latest/"),
+    Package(pypi="cyclopts", index_url="https://cyclopts.readthedocs.io/en/latest/"),
     # ENTRY-LINE-MARKER
 ]
 # fmt: on