+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+Nothing noticeable unreleased.
+
+## v2025.6
+
 - Added a `-p/--print` flag to the `pyrefdev` CLI to print the URL instead of opening it in the browser.
 - Support [cyclopts](https://cyclopts.readthedocs.io/en/latest/).
 - Support [yib](https://yib.readthedocs.io/en/latest/).